requirements.txt
flask 
pillow 
numpy
